+++ lib/src/user_agent_application.dart
@@ -95,6 +95,19 @@ class UserAgentApplication {
     return AuthResponse._fromJsObject(response);
   }
 
+  /// Establishes user context silently with SSO when an existing session is known.
+  /// 
+  /// The [request] must provide a `loginHint` or `sid`.
+  /// 
+  /// See https://github.com/AzureAD/microsoft-authentication-library-for-js/tree/dev/lib/msal-core#ssosilent
+  /// for more information.
+  Future<AuthResponse> ssoSilent(AuthRequest request) async {
+    final response = await _convertMsalPromise<interop.AuthResponse>(
+        _callJsMethod(() => _jsObject.ssoSilent(request._jsObject)));
+
+    return AuthResponse._fromJsObject(response);
+  }
+
   /// Gets all currently cached unique accounts based on `homeAccountIdentifier`.
   List<Account> getAllAccounts() {
     return _callJsMethod(() => _jsObject