+++ test/user_agent_application_test.dart
@@ -45,6 +45,15 @@ void main() {
         throwsA(isA<ClientAuthException>()));
   });
 
+  test('ssoSilent converts JS errors', () {
+    final userAgentApp = UserAgentApplication(Configuration());
+
+    expect(
+        () => userAgentApp
+            .ssoSilent(AuthRequest()..scopes = ['user.read']),
+        throwsA(isA<ClientAuthException>()));
+  });
+
   test('getLoginInProgress works', () {
     final userAgentApp = UserAgentApplication(Configuration());