+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## v1.3.0
+- Added `UserAgentApplication.ssoSilent`.
 - Added `UserAgentApplication.setLogger`.
 - Added `AuthRequest.redirectStartPage`.
 - Fixed exception thrown when creating a `Logger` without any options.